perf: split user model to many file
